+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [3.2.1] - 2021-11-05
+
+### Fixed
+- Bump glob-parent from 5.0.0 to 5.1.2
+- Bump axios from 0.19.2 to 0.21.2
+- Bump lodash from 4.17.15 to 4.17.21
 
 ## [3.2.0] - 2021-07-14
 ### Changed